Add basic list functionality to chatbot
Add ArrayList field to Friday which will store the inputs of the user. Add printlist() method to Friday which handles the logic for generating the string responsible for displaying the stored user inputs when user prompts the bot with "list". Adjusted start() logic to handle the prompt "list". Added additional String formatting to handleInput() in UserInputHandler.java.