Create Task class
Task class encapsulates the user input by storing the task description and an additional boolean field to determine the status of the Task. Add markAsDone() and markAsUndone() methods to change the isDone state. Change ArrayList in Friday to store Task objects instead. Add mark and unmark user input logic to start() in Friday.