Command class and command subclasses
Added further abstraction by encapsulating user input processing to an abstract command class. Commands can be executed using execute(), which handles the logic to add a task to the TaskList, handles exceptions if any and also returns a ui output via the terminal. This simplifies the Parser class, increasing readability.