Priority command
Added priority command to Friday's available commands. Every task now contains a field which indicates the priority of a task. H stands for high, M stands for medium and L stands for low.

The priority command requires an index (according to the index shown in the TaskList) and a relevant priority field after the /lvl command to assign a priority to a task.