GUI improvements to Friday
Made overall improvements to the Friday GUI.

Changed the Friday and User icons to circular icons to
make them more appealing. Added a title to the GUI as
requested by the module to differentiate from Duke. Also
added promptText to the userInput bar to help guide users
to interact with it. Improved the dialog box by introducing
text bubbles with margins and borders. Additionally, styled
the send button and added a background to the chat.