Refactor Friday Commands
Most of the Friday Commands experience repeat code due to having to generate both
a GUI response and CLI response. The logic is mostly consistent but each response
require a different return value.

Refactored most of the Friday Commands to reduce this repetitiveness. Additionally,
utilised SLAP to improve code quality.